Allow cluster name to be dynamic after unset on reload
In case the cluster name is removed on a config reload,
the members of that cluster should renegotiate a new
dynamic cluster name in case an explicit cluster name
was not set.

This also adds logic so that in case membership from
a cluster has changed after the reload, the server
disconnects the routes that are no longer considered
to form part of the same cluster.

Signed-off-by: Waldemar Quevedo <wally@synadia.com>